# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FasalCast / KhetiSetu API",
    "description": "Farmer recommendation, business dashboard, and RAG endpoints for KhetiSetu.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/districts": {
      "get": {
        "tags": [
          "reference"
        ],
        "summary": "List Districts",
        "operationId": "list_districts_api_v1_districts_get",
        "parameters": [
          {
            "name": "state_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "State Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DistrictOut"
                  },
                  "title": "Response List Districts Api V1 Districts Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/products": {
      "get": {
        "tags": [
          "reference"
        ],
        "summary": "List Products",
        "operationId": "list_products_api_v1_products_get",
        "parameters": [
          {
            "name": "fertilizer_type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Fertilizer Type"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProductOut"
                  },
                  "title": "Response List Products Api V1 Products Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/business/dashboard": {
      "get": {
        "tags": [
          "business"
        ],
        "summary": "Dashboard",
        "operationId": "dashboard_api_v1_business_dashboard_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "district_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "District Id"
            }
          },
          {
            "name": "season_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Season Id"
            }
          },
          {
            "name": "year",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Year"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BusinessDashboardResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/business/forecast": {
      "get": {
        "tags": [
          "business"
        ],
        "summary": "Forecast",
        "operationId": "forecast_api_v1_business_forecast_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "district_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "District Id"
            }
          },
          {
            "name": "product_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Product Id"
            }
          },
          {
            "name": "year",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Year"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ForecastItem"
                  },
                  "title": "Response Forecast Api V1 Business Forecast Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/business/inventory": {
      "get": {
        "tags": [
          "business"
        ],
        "summary": "Inventory",
        "operationId": "inventory_api_v1_business_inventory_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "district_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "District Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/InventoryItem"
                  },
                  "title": "Response Inventory Api V1 Business Inventory Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/business/transfers": {
      "get": {
        "tags": [
          "business"
        ],
        "summary": "Transfers",
        "operationId": "transfers_api_v1_business_transfers_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/TransferItem"
                  },
                  "type": "array",
                  "title": "Response Transfers Api V1 Business Transfers Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/business/alerts": {
      "get": {
        "tags": [
          "business"
        ],
        "summary": "Alerts",
        "operationId": "alerts_api_v1_business_alerts_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "district_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "District Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AlertItem"
                  },
                  "title": "Response Alerts Api V1 Business Alerts Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AlertItem": {
        "properties": {
          "district": {
            "type": "string",
            "title": "District"
          },
          "product": {
            "type": "string",
            "title": "Product"
          },
          "severity": {
            "type": "string",
            "title": "Severity"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "district",
          "product",
          "severity",
          "message"
        ],
        "title": "AlertItem"
      },
      "BusinessDashboardResponse": {
        "properties": {
          "season": {
            "type": "string",
            "title": "Season"
          },
          "expected_input_demand": {
            "items": {
              "$ref": "#/components/schemas/InputDemandItem"
            },
            "type": "array",
            "title": "Expected Input Demand"
          },
          "farmer_crop_intent": {
            "items": {
              "$ref": "#/components/schemas/CropIntentSummaryItem"
            },
            "type": "array",
            "title": "Farmer Crop Intent"
          },
          "alerts": {
            "items": {
              "$ref": "#/components/schemas/AlertItem"
            },
            "type": "array",
            "title": "Alerts"
          },
          "recommended_action": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RecommendedAction"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "season",
          "expected_input_demand",
          "farmer_crop_intent",
          "alerts",
          "recommended_action"
        ],
        "title": "BusinessDashboardResponse"
      },
...
      "CropIntentSummaryItem": {
        "properties": {
          "crop": {
            "type": "string",
            "title": "Crop"
          },
          "acres": {
            "type": "number",
            "title": "Acres"
          }
        },
        "type": "object",
        "required": [
          "crop",
          "acres"
        ],
        "title": "CropIntentSummaryItem"
      },
...
      "DistrictOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "state_id": {
            "type": "integer",
            "title": "State Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "latitude": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latitude"
          },
          "longitude": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Longitude"
          },
          "also_known_as": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Also Known As"
          }
        },
        "type": "object",
        "required": [
          "id",
          "state_id",
          "name",
          "latitude",
          "longitude",
          "also_known_as"
        ],
        "title": "DistrictOut"
      },
...
      "ForecastItem": {
        "properties": {
          "product_id": {
            "type": "integer",
            "title": "Product Id"
          },
          "year": {
            "type": "integer",
            "title": "Year"
          },
          "month": {
            "type": "integer",
            "title": "Month"
          },
          "predicted_demand": {
            "type": "number",
            "title": "Predicted Demand"
          },
          "lower_bound": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lower Bound"
          },
          "upper_bound": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Upper Bound"
          },
          "confidence": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confidence"
          },
          "model_version": {
            "type": "string",
            "title": "Model Version"
          }
        },
        "type": "object",
        "required": [
          "product_id",
          "year",
          "month",
          "predicted_demand",
          "lower_bound",
          "upper_bound",
          "confidence",
          "model_version"
        ],
        "title": "ForecastItem"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "InputDemandItem": {
        "properties": {
          "product": {
            "type": "string",
            "title": "Product"
          },
          "quantity_mt": {
            "type": "number",
            "title": "Quantity Mt"
          }
        },
        "type": "object",
        "required": [
          "product",
          "quantity_mt"
        ],
        "title": "InputDemandItem"
      },
      "InventoryItem": {
        "properties": {
          "product_id": {
            "type": "integer",
            "title": "Product Id"
          },
          "quantity": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Quantity"
          },
          "batch_no": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Batch No"
          },
          "manufacturing_date": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Manufacturing Date"
          },
          "expiry_date": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expiry Date"
          }
        },
        "type": "object",
        "required": [
          "product_id",
          "quantity",
          "batch_no",
          "manufacturing_date",
          "expiry_date"
        ],
        "title": "InventoryItem"
      },
...
      "ProductOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "product_name": {
            "type": "string",
            "title": "Product Name"
          },
          "product_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Product Type"
          },
          "fertilizer_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fertilizer Type"
          }
        },
        "type": "object",
        "required": [
          "id",
          "product_name",
          "product_type",
          "fertilizer_type"
        ],
        "title": "ProductOut"
      },
...
      "RecommendedAction": {
        "properties": {
          "product": {
            "type": "string",
            "title": "Product"
          },
          "forecast_mt": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Forecast Mt"
          },
          "current_stock_mt": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Stock Mt"
          },
          "safety_stock_mt": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Safety Stock Mt"
          },
          "recommended_dispatch_mt": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Recommended Dispatch Mt"
          },
          "action": {
            "type": "string",
            "title": "Action"
          }
        },
        "type": "object",
        "required": [
          "product",
          "forecast_mt",
          "current_stock_mt",
          "safety_stock_mt",
          "recommended_dispatch_mt",
          "action"
        ],
        "title": "RecommendedAction"
      },
...
      "TransferItem": {
        "properties": {
          "product_id": {
            "type": "integer",
            "title": "Product Id"
          },
          "from_district_id": {
            "type": "integer",
            "title": "From District Id"
          },
          "to_district_id": {
            "type": "integer",
            "title": "To District Id"
          },
          "recommended_transfer_qty": {
            "type": "number",
            "title": "Recommended Transfer Qty"
          },
          "reason": {
            "type": "string",
            "title": "Reason"
          }
        },
        "type": "object",
        "required": [
          "product_id",
          "from_district_id",
          "to_district_id",
          "recommended_transfer_qty",
          "reason"
        ],
        "title": "TransferItem"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}